Move plugin registry under the pkg/app/pipedv1/plugin
